add options for simd and binary-search
